Update modify.py
This program modifies the parsed files by adding their respective meta data from a version of theirs that had not been parsed. There are hundreds of such files, so this code scans two folders (target folder containing all the parsed files and a source folder containing files of the same name that are not parsed but containing meta data that the final version needs) to select files of the same name and do the necessary changes.
This code runs on the coli servers at Saarland University. One can reuse the code for their specific tasks just by changing the required path.